syntax corrections
creates login tests
