+++ src/test/java/testcases/AllSongsTests.java
@@ -46,7 +46,6 @@ public class AllSongsTests extends BaseTest {
     @Test(description = "Click on the Album Tab in the Info Panel")
     public void clickInfoPanelAlbumTab() {
         loginPage.loginValidCredentials().clickAllSongs();
-
         allSongsPage
                 .checkHeaderTitle()
                 .contextClickFirstSong()
@@ -57,14 +56,12 @@ public class AllSongsTests extends BaseTest {
     @Test(description="Like all songs")
     public void likeAll(){
         loginPage.loginValidCredentials().clickAllSongs();
-
         allSongsPage.likeSongs();
         Assert.assertFalse(allSongsPage.checkUnliked());
     }
     @Test(description="Unlike all liked songs", dependsOnMethods = {"likeAll"})
     public void unlikeAll () {
         loginPage.loginValidCredentials().clickAllSongs();
-
         allSongsPage.unlikeSongs();
             Assert.assertTrue(allSongsPage.checkUnliked());
     }
